Added_Import_Module_Connect_AZAccount
When running locally it is required to connect to an Azure account (through the Azure module). If you mention "yes" or "y" it will import the module and connect to Azure. If you say anything else or just skip the question, it will not import any module (as you may execute from cloudshell)